+++ matthewplotlib/__init__.py
@@ -58,6 +58,7 @@ from matthewplotlib.plots import (
     histogram,
     columns,
     vistogram,
+    candles,
     hilbert,
     calendar,
     weeks,